+++ src/main/java/terracraft/mixin/item/metaldetector/client/GuiMixin.java
@@ -4,6 +4,7 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.Gui;
+import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
@@ -15,6 +16,8 @@ import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import terracraft.TerraCraft;
+import terracraft.common.init.ModBlocks;
 import terracraft.common.init.ModItems;
 import terracraft.common.trinkets.TrinketsHelper;
 
@@ -24,14 +27,15 @@ import java.util.stream.Stream;
 @Mixin(Gui.class)
 public abstract class GuiMixin {
 
-	@Shadow @Final private Minecraft minecraft;
 	@Shadow private int screenHeight;
 	@Shadow private int screenWidth;
 	@Unique private int timer = 0;
-	@Unique private String lastOre = getNearRareOre();
+	@Unique TranslatableComponent detectedText = new TranslatableComponent(TerraCraft.MOD_ID + ".ui.oreDetected");
+	@Unique TranslatableComponent noDetectedText = new TranslatableComponent(TerraCraft.MOD_ID + ".ui.oreNotDetected");
+	@Unique private String lastOre;
 
 	@Shadow protected abstract Player getCameraPlayer();
-	@Shadow protected abstract Font getFont();
+	@Shadow public abstract Font getFont();
 
 	@Inject(method = "renderPlayerHealth", require = 0, at = @At(value = "TAIL"))
 	private void renderGuiClock(PoseStack matrices, CallbackInfo ci) {
@@ -39,6 +43,9 @@ public abstract class GuiMixin {
 
 		if (player == null || !getEquippedTrinkets(player)) {
 			return;
+		}
+		if (lastOre == null) {
+			lastOre = noDetectedText.getString();
 		}
 
 		timer += 1;
@@ -59,12 +66,8 @@ public abstract class GuiMixin {
 	private boolean getEquippedTrinkets(Player player) {
 		boolean equipped = false;
 
-		if (TrinketsHelper.isEquipped(ModItems.METAL_DETECTOR, player) || TrinketsHelper.isEquipped(ModItems.GOBLIN_TECH, player) || TrinketsHelper.isEquipped(ModItems.PDA, player)
-				|| TrinketsHelper.isEquipped(ModItems.CELL_PHONE, player)) {
-			equipped = true;
-		}
-		if (player.getInventory().contains(ModItems.METAL_DETECTOR.getDefaultInstance()) || player.getInventory().contains(ModItems.GOBLIN_TECH.getDefaultInstance()) || player.getInventory().contains(ModItems.PDA.getDefaultInstance())
-				|| player.getInventory().contains(ModItems.CELL_PHONE.getDefaultInstance())) {
+		if (TrinketsHelper.isInInventory(ModItems.METAL_DETECTOR, player) || TrinketsHelper.isInInventory(ModItems.GOBLIN_TECH, player) || TrinketsHelper.isInInventory(ModItems.PDA, player)
+				|| TrinketsHelper.isInInventory(ModItems.CELL_PHONE, player)) {
 			equipped = true;
 		}
 
@@ -79,6 +82,7 @@ public abstract class GuiMixin {
 		boolean emeraldOre = false;
 		boolean diamondOre = false;
 		boolean goldOre = false;
+		boolean demoniteOre = false;
 		boolean lapisOre = false;
 		boolean redstoneOre = false;
 		boolean ironOre = false;
@@ -86,7 +90,7 @@ public abstract class GuiMixin {
 		boolean coalOre = false;
 		boolean netherQuartz = false;
 		if (mc.player != null && mc.level != null) {
-			AABB test = new AABB(
+			AABB detectionBounds = new AABB(
 					mc.player.position().x() - 7.5f,
 					mc.player.position().y() - 7.5f,
 					mc.player.position().z() - 7.5f,
@@ -94,7 +98,7 @@ public abstract class GuiMixin {
 					mc.player.position().y() + 7.5f,
 					mc.player.position().z() + 7.5f
 			);
-			Stream<BlockState> blocks = mc.level.getBlockStates(test);
+			Stream<BlockState> blocks = mc.level.getBlockStates(detectionBounds);
 			List<BlockState> blocksList = blocks.toList();
 			for (BlockState block : blocksList) {
 				if (block.getBlock().equals(Blocks.ANCIENT_DEBRIS)) {
@@ -105,6 +109,8 @@ public abstract class GuiMixin {
 					diamondOre = true;
 				} else if (block.getBlock().equals(Blocks.GOLD_ORE) || block.getBlock().equals(Blocks.DEEPSLATE_GOLD_ORE) || block.getBlock().equals(Blocks.NETHER_GOLD_ORE)) {
 					goldOre = true;
+				} else if (block.getBlock().equals(ModBlocks.DEMONITE_ORE) || block.getBlock().equals(ModBlocks.DEEPSLATE_DEMONITE_ORE)) {
+					demoniteOre = true;
 				} else if (block.getBlock().equals(Blocks.LAPIS_ORE) || block.getBlock().equals(Blocks.DEEPSLATE_LAPIS_ORE)) {
 					lapisOre = true;
 				} else if (block.getBlock().equals(Blocks.REDSTONE_ORE) || block.getBlock().equals(Blocks.DEEPSLATE_REDSTONE_ORE)) {
@@ -121,27 +127,29 @@ public abstract class GuiMixin {
 			}
 		}
 		if (ancientDebris) {
-			sb.append("Ancient Debris Detected Nearby");
+			sb.append(Blocks.ANCIENT_DEBRIS.getName().getString()).append(" ").append(detectedText.getString());
 		} else if (emeraldOre) {
-			sb.append("Emerald Ore Detected Nearby");
+			sb.append(Blocks.EMERALD_ORE.getName().getString()).append(" ").append(detectedText.getString());
 		} else if (diamondOre) {
-			sb.append("Diamond Ore Detected Nearby");
+			sb.append(Blocks.DIAMOND_ORE.getName().getString()).append(" ").append(detectedText.getString());
 		} else if (goldOre) {
-			sb.append("Gold Ore Detected Nearby");
+			sb.append(Blocks.GOLD_ORE.getName().getString()).append(" ").append(detectedText.getString());
+		} else if (demoniteOre) {
+			sb.append(ModBlocks.DEMONITE_ORE.getName().getString()).append(" ").append(detectedText.getString());
 		} else if (lapisOre) {
-			sb.append("Lapis Ore Detected Nearby");
+			sb.append(Blocks.LAPIS_ORE.getName().getString()).append(" ").append(detectedText.getString());
 		} else if (redstoneOre) {
-			sb.append("Redstone Ore Detected Nearby");
+			sb.append(Blocks.REDSTONE_ORE.getName().getString()).append(" ").append(detectedText.getString());
 		} else if (ironOre) {
-			sb.append("Iron Ore Detected Nearby");
+			sb.append(Blocks.IRON_ORE.getName().getString()).append(" ").append(detectedText.getString());
 		} else if (copperOre) {
-			sb.append("Copper Ore Detected Nearby");
+			sb.append(Blocks.COPPER_ORE.getName().getString()).append(" ").append(detectedText.getString());
 		} else if (coalOre) {
-			sb.append("Coal Ore Detected Nearby");
+			sb.append(Blocks.COAL_ORE.getName().getString()).append(" ").append(detectedText.getString());
 		} else if (netherQuartz) {
-			sb.append("Nether Quartz Detected Nearby");
+			sb.append(Blocks.NETHER_QUARTZ_ORE.getName().getString()).append(" ").append(detectedText.getString());
 		} else {
-			sb.append("No Ores Detected Nearby");
+			sb.append(noDetectedText.getString());
 		}
 		return sb.toString();
 	}